cv-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pillow==10.4.0
numpy==2.0.2
ultralytics==8.3.29
python-multipart==0.0.9

